added stylecop

+++ BLL/Services/Implementations/UserService.cs
@@ -1,10 +1,10 @@
-﻿using BLL.Models;
+﻿using System.Linq.Expressions;
+using System.Text.RegularExpressions;
+using BLL.Models;
 using BLL.Services.Interfaces;
 using BLL.Utils;
 using DAL.Models;
 using DAL.Repositories;
-using System.Linq.Expressions;
-using System.Text.RegularExpressions;
 
 namespace BLL.Services.Implementations
 {
@@ -34,14 +34,14 @@ namespace BLL.Services.Implementations
                     Password = user.Password,
                     CreatedDate = DateTime.UtcNow,
                     ModifiedDate = DateTime.UtcNow,
-                    Role = UserRole.User
+                    Role = UserRole.User,
                 };
 
                 await _userRepo.AddAsync(dalUser);
 
                 dalUser = await _userRepo.GetOneAsync(u => u.NickName == user.NickName);
 
-                return new Result<UserModel>(registerResult.IsError, "Register succeded.", DALUserToBLLUser(dalUser));
+                return new Result<UserModel>(registerResult.IsError, "Register succeded.", DALUserToBLLUser(dalUser!));
             }
 
             return registerResult;
@@ -62,9 +62,13 @@ namespace BLL.Services.Implementations
         {
             var users = await _userRepo.GetAllAsync();
             var bllUsers = new List<UserModel>();
-            foreach(var user in users)
+            foreach (var user in users)
             {
-                bllUsers.Add(DALUserToBLLUser(user));
+                var bllUser = DALUserToBLLUser(user);
+                if (bllUser != null)
+                {
+                    bllUsers.Add(bllUser);
+                }
             }
 
             return bllUsers;
@@ -86,9 +90,12 @@ namespace BLL.Services.Implementations
         public async Task UpdateUser(UserModel user)
         {
             var dalUser = await BLLUserToDALUserAsync(user);
-            await _userRepo.Update(dalUser);
+            if (dalUser != null)
+            {
+                await _userRepo.Update(dalUser);
+            }
         }
-        
+
         public async Task DeleteUserAsync(Guid id)
         {
             await _userRepo.DeleteAsync(id);
@@ -97,23 +104,52 @@ namespace BLL.Services.Implementations
         public async Task DeleteUser(UserModel user)
         {
             var dalUser = await BLLUserToDALUserAsync(user);
-            await _userRepo.Delete(dalUser);
+            if (dalUser != null)
+            {
+                await _userRepo.Delete(dalUser);
+            }
+        }
+
+        public async Task<bool> ChatExists(Guid user1, Guid user2)
+        {
+            var user1Dal = await _userRepo.GetByIdAsync(user1);
+            if (user1Dal!.Chats == null)
+            {
+                return false;
+            }
+
+            return user1Dal.Chats.Any(chat => chat.Users.Any(user => user.Id == user2));
+        }
+
+        public async Task<User?> BLLUserToDALUserAsync(UserModel user)
+        {
+            return await _userRepo.GetByIdAsync(user.Id);
+        }
+
+        public static UserModel? DALUserToBLLUser(User user)
+        {
+            if (user != null)
+            {
+                return new UserModel(user.Id, user.NickName!, user.FirstName!, user.LastName!, user.Email!, user.Password!);
+            }
+
+            return null;
         }
 
         private async Task<Result<UserModel>> IsValidRegistrationData(UserModel user)
         {
-            string message = "";
+            string message = string.Empty;
             bool passwordValidity = user.Password != null
                 && new Regex("^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{8,}$").IsMatch(user.Password);
-            message += passwordValidity ? "" : "Invalid password.\t";
+            message += passwordValidity ? string.Empty : "Invalid password.\t";
 
             bool emailValidity = user.Email != null && new Regex("^\\S+@\\S+\\.\\S+$").IsMatch(user.Email);
-            message += emailValidity ? "" : "Invalid email.\t";
+            message += emailValidity ? string.Empty : "Invalid email.\t";
 
             bool requiredFieldsPresent = user.NickName != null && user.FirstName != null && user.LastName != null;
-            message += requiredFieldsPresent ? "" : "Required fields are missing.";
+            message += requiredFieldsPresent ? string.Empty : "Required fields are missing.";
 
-            User existingUser = await UserExists(user.NickName);
+            User? existingUser = await UserExists(user.NickName!);
             if (existingUser != null)
             {
                 return new Result<UserModel>(true, "User with provided nickname already exists");
@@ -127,16 +163,17 @@ namespace BLL.Services.Implementations
             if (user.Password != null && user.NickName != null)
             {
                 var password = user.Password;
-                User existingUser = await UserExists(user.NickName);
+                User? existingUser = await UserExists(user.NickName!);
                 if (existingUser == null)
                 {
                     return new Result<UserModel>(true, "User doesn`t exist.");
                 }
-                bool res = PasswordUtil.IsValidPassword(password, existingUser.Password);
+
+                bool res = PasswordUtil.IsValidPassword(password, existingUser.Password!);
                 if (res)
                 {
                     var bllUser = DALUserToBLLUser(existingUser);
-                    return new Result<UserModel>(!res, "Login Success", bllUser);
+                    return new Result<UserModel>(!res, "Login Success", bllUser!);
                 }
 
                 return new Result<UserModel>(true, "Invalid Password");
@@ -145,36 +182,10 @@ namespace BLL.Services.Implementations
             return new Result<UserModel>(true, "Nickname and password field were not provided");
         }
 
-        public async Task<bool> ChatExists(Guid user1, Guid user2)
-        {
-            var user1Dal = await _userRepo.GetByIdAsync(user1);
-            if(user1Dal.Chats == null) 
-                return false;
-
-            return user1Dal.Chats.Any(chat => chat.Users.Any(user => user.Id == user2));
-        }
-        
-      
-        public async Task<User?> BLLUserToDALUserAsync(UserModel user)
-        {
-            return await _userRepo.GetByIdAsync(user.Id);
-        }
-
-        public static UserModel? DALUserToBLLUser(User user)
-        {
-            if (user != null)
-            {
-                return new UserModel(user.Id, user.NickName, user.FirstName, user.LastName, user.Email, user.Password);
-            }
-
-            return null;
-        }
-
         private async Task<User?> UserExists(string nickname)
         {
             var userData = await _userRepo.GetOneAsync(filter: u => u.NickName == nickname);
             return userData;
-
         }
     }
 }
